Add rudimentary support for thumbsticks
PS-One style thumbsticks (the ones found on nearly
every console controller since the PS-One).

The L3/R3 button is pressed if and only if the first
finger exerts more pressure than a threshold (for the
moment, the average value of the minimum and maximum
force).

The non-pressure-sensitive alternative is to consider
the L3/R3 button pressed if and only if there is more
than one finger on the same joystick.