Set fixed values for pressure detection
This commit removes the prompt to calibrate the screen, and only
checks if the screen can detect pressure at all. For the moment,
the calibration values are stored as constants in the JS file.

This will be easier to customise for the user in the future, but
having the prompt every time the controller is used was more
trouble than it was worth.