Yoke can now read numeric variables from CSS
This commit adds CSS variables and a codebase that for now, is unused.
In a future commit, most variables will be read from the CSS instead of
hardcoded in the JavaScript.
This allows users to both fine-tune Yoke more easily, and also to fix
settings per layout and per control (for example, the first button from
a given layout might need extra force to be pushed).

CSS blocks have also been moved around for the user's convenience.
Critical parameters that users should not touch are now at the bottom of
the file.
Parameters at the top will be overridden by parameters at the bottom,
following normal CSS priority rules.
No examples are given for this, but parameters bound to any specific
control by ID (e.g. "#j1") would be given a higher priority.