Add support for joystick shapes and deadzones
Joysticks can now be elliptical, square, round or rectangular.
A central deadzone is also supported, which depends on the joystick shape:
rectangular joysticks have axial deadzones (evaluated separately for each
component), while round joysticks have a single radial deadzone.

Both deadzones are scaled, meaning that output grows gradually from the
edge of the deadzone to the edge of the joystick.