CChange sorting method for controls
Several programs on Linux (and all programs in Windows, due to a design
decision on Yoke) ignore the semantics of every button and only look at
the order in which they are reported.

Because of this, the JavaScript side now reorders controls to simplify
most configurations (face buttons should go first, then in-game menu
buttons, then thumbstick buttons, and then special buttons like the
branded key to open the system menu, which is even less used in this
scenario).